networks to `consts`, separate network `brand` to `assets`
