+++ packages/app/src/api/index.ts
@@ -1,7 +1,7 @@
 // Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import { NetworkList, SystemChainList } from 'config/networks'
+import { NetworkList, SystemChainList } from 'consts/networks'
 import { Subscriptions } from 'controllers/Subscriptions'
 import type { PolkadotClient } from 'polkadot-api'
 import { createClient } from 'polkadot-api'
