+++ packages/app/src/api/util.ts
@@ -3,7 +3,7 @@
 
 import type { ApiChainType } from 'api/types'
 import type { AnyApi } from 'common-types'
-import { NetworkList, SystemChainList } from 'config/networks'
+import { NetworkList, SystemChainList } from 'consts/networks'
 import type { ChainId } from 'types'
 
 // Get the light client metadata for the given chain type and network.
