+++ packages/app/src/config/pages.ts
@@ -9,7 +9,6 @@ import { Pools } from 'pages/Pools'
 import { Rewards } from 'pages/Rewards'
 import { Validators } from 'pages/Validators'
 
-const BASE_URL = import.meta.env.BASE_URL
 export const PageCategories: PageCategoryItems = [
   {
     id: 1,
@@ -29,7 +28,7 @@ export const PagesConfig: PagesConfigItems = [
   {
     category: 2,
     key: 'overview',
-    uri: `${BASE_URL}`,
+    uri: `${import.meta.env.BASE_URL}`,
     hash: '/overview',
     Entry: Overview,
     lottie: 'globe',
@@ -37,7 +36,7 @@ export const PagesConfig: PagesConfigItems = [
   {
     category: 2,
     key: 'pools',
-    uri: `${BASE_URL}pools`,
+    uri: `${import.meta.env.BASE_URL}pools`,
     hash: '/pools',
     Entry: Pools,
     lottie: 'groups',
@@ -45,7 +44,7 @@ export const PagesConfig: PagesConfigItems = [
   {
     category: 2,
     key: 'nominate',
-    uri: `${BASE_URL}nominate`,
+    uri: `${import.meta.env.BASE_URL}nominate`,
     hash: '/nominate',
     Entry: Nominate,
     lottie: 'trending',
@@ -53,7 +52,7 @@ export const PagesConfig: PagesConfigItems = [
   {
     category: 2,
     key: 'rewards',
-    uri: `${BASE_URL}rewards`,
+    uri: `${import.meta.env.BASE_URL}rewards`,
     hash: '/rewards',
     Entry: Rewards,
     lottie: 'analytics',
@@ -61,7 +60,7 @@ export const PagesConfig: PagesConfigItems = [
   {
     category: 3,
     key: 'validators',
-    uri: `${BASE_URL}validators`,
+    uri: `${import.meta.env.BASE_URL}validators`,
     hash: '/validators',
     Entry: Validators,
     lottie: 'view',
@@ -69,7 +68,7 @@ export const PagesConfig: PagesConfigItems = [
   {
     category: 3,
     key: 'operators',
-    uri: `${BASE_URL}operators`,
+    uri: `${import.meta.env.BASE_URL}operators`,
     hash: '/operators',
     Entry: Operators,
     lottie: 'label',
