+++ packages/app/src/contexts/Network/types.ts
@@ -1,8 +1,7 @@
 // Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { Network } from 'common-types'
-import type { NetworkId } from 'types'
+import type { Network, NetworkId } from 'types'
 
 export interface NetworkState {
   name: NetworkId
