+++ packages/app/src/library/Headers/Popovers/MenuPopover.tsx
@@ -11,8 +11,8 @@ import {
 } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { useOutsideAlerter } from '@w3ux/hooks'
-import LanguageSVG from 'assets/svg/icons/language.svg?react'
-import MoonOutlineSVG from 'assets/svg/icons/moon.svg?react'
+import LanguageSVG from 'assets/icons/language.svg?react'
+import MoonOutlineSVG from 'assets/icons/moon.svg?react'
 import { GitHubURl } from 'consts'
 import { useCurrency } from 'contexts/Currency'
 import { usePlugins } from 'contexts/Plugins'
