+++ packages/app/src/overlay/modals/MailSupport/index.tsx
@@ -1,7 +1,7 @@
 // Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import MailSVG from 'assets/svg/icons/mail.svg?react'
+import MailSVG from 'assets/icons/mail.svg?react'
 import { MailSupportAddress } from 'consts'
 import { Title } from 'library/Modal/Title'
 import { useTranslation } from 'react-i18next'
