+++ packages/app/src/pages/Overview/AccountBalance/BalanceChart.tsx
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import { faCheck, faCheckDouble } from '@fortawesome/free-solid-svg-icons'
+import { getChainIcons } from 'assets'
 import BigNumber from 'bignumber.js'
 import { useActiveAccounts } from 'contexts/ActiveAccounts'
 import { useBalances } from 'contexts/Balances'
@@ -23,11 +24,8 @@ import { planckToUnitBn } from 'utils'
 export const BalanceChart = () => {
   const { t } = useTranslation('pages')
   const {
-    networkData: {
-      units,
-      unit,
-      brand: { token: Token },
-    },
+    network,
+    networkData: { units, unit },
   } = useNetwork()
   const { currency } = useCurrency()
   const { openModal } = useOverlay().modal
@@ -37,6 +35,7 @@ export const BalanceChart = () => {
   const { accountHasSigner } = useImportedAccounts()
   const { feeReserve, getTransferOptions } = useTransferOptions()
 
+  const Token = getChainIcons(network).token
   const balance = getBalance(activeAccount)
   const allTransferOptions = getTransferOptions(activeAccount)
   const { edReserved } = allTransferOptions
