+++ packages/consts/package.json
@@ -11,6 +11,10 @@
     "clear": "rm -rf build tsconfig.tsbuildinfo dist"
   },
   "dependencies": {
+    "assets": "workspace:*",
+    "polkadot-api": "^1.9.5"
+  },
+  "devDependencies": {
     "types": "workspace:*"
   }
 }
