+++ packages/types/src/networks.ts
@@ -1,6 +1,8 @@
 // Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
+import type { FunctionComponent, SVGProps } from 'react'
+
 export type ChainId = NetworkId | SystemChainId
 
 export type NetworkId = 'polkadot' | 'kusama' | 'westend'
@@ -9,3 +11,49 @@ export type SystemChainId =
   | 'people-polkadot'
   | 'people-kusama'
   | 'people-westend'
+
+export type Networks = Record<string, Network>
+
+export interface Network {
+  name: NetworkId
+  endpoints: {
+    lightClientKey: string
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    lightClient: () => Promise<any>
+    defaultRpcEndpoint: string
+    rpcEndpoints: Record<string, string>
+  }
+  unit: string
+  units: number
+  ss58: number
+  defaultFeeReserve: number
+}
+
+export interface SystemChain {
+  name: string
+  ss58: number
+  units: number
+  unit: string
+  endpoints: {
+    lightClientKey: string
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    lightClient: () => Promise<any>
+    rpcEndpoints: Record<string, string>
+  }
+  relayChain: NetworkId
+}
+
+export interface ChainIcons {
+  icon: FunctionComponent<
+    SVGProps<SVGSVGElement> & { title?: string | undefined }
+  >
+  token: FunctionComponent<
+    SVGProps<SVGSVGElement> & { title?: string | undefined }
+  >
+  inline: {
+    svg: FunctionComponent<
+      SVGProps<SVGSVGElement> & { title?: string | undefined }
+    >
+    size: string
+  }
+}
